fixed .xcodeproj not getting nse dependency
(This commit isn't intented be merged, should be brokend up)

The issue is `PBXTargetDependency` and `PBXTargetDependency` are not in
xcodeproj files by default and the `cordova-node-xcode` is not
handling this correctly with their `addTarget` and `addTargetDependency`
API.
A workaround was added to set these to a blank object so
`addTargetDependency` will corretly add them.

addBuildPhase([], "PBXTargetDependency",..) did not work as
this as extra invalid properties the breaks the .xcodeproj.