Refactor `NseUpdaterManager`
Motivation: this change accomplishes two main things:
1) write-safety by ensuring the files are modified one at a time by the updater functions (make synchronous)
2) Converts the manager to an instantiated class that takes the iOS directory path

Previous behavior: we were modifying the template files directly (in the node_module) and *then* copying them over to the xcode project. Why is this a problem? Because once these files' replacement target strings were modified with dynamic values, re-running the build would not update them correctly (since they were no longer the template replacement strings).

New behavior: we copy the template files to the xcode project and *only then* modify them. To do this, we need the iOS directory path which is passed in via the constructor.

Note: because of the linting rules, we must catch the promises.