Check that plugin props are being provided
Motivation: the property "mode" is required. Currently, because we don't check for this, nodejs is throwing a runtime error when trying to access smallIcons (optional prop) when the props object is undefined.

This is confusing for folks who first set up and try running without setting a props object to the plugin array.

Clear error message will be much more helpful.